refactor: 회원 관련 패키지 생성
